0316 Java Lesson & PolyMorphism practice example

+++ JavaClass/src/daySixth/Game/Thing.java
@@ -1,6 +1,6 @@
 package daySixth.Game;
 
 public interface Thing {
-    int getWeight();
-    void setWeight(int weight);
+    double getWeight();
+    void setWeight(double weight);
 }
